Fixing typos and syntax from User Manual
There was emphasising and bolding instructions in wrong order when
combined. At least on my system this must be started with *+ and ended
with +*. (There was also two cases when the ending was in same
order as starting also spoiling the XML syntax.) These prevented XML
validation thus terminating the compilation of the final documentation.
I fixed also a bunch of typos while at it.

Signed-off-by: Miika Turkia <miika.turkia@gmail.com>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>